pyproject.toml
[tool.poetry]
name = "air-accountant"
version = "0.1.0"
description = "An AI assistant for accounting using Airtable as database."
authors = ["deejungx <dipzz1394@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyairtable = "^2.2.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
